+++ test/test_helper.rb
@@ -7,8 +7,6 @@ require 'active_support'
 require 'test/spec'
 require 'mocha'
 
-Workling.try_load_a_memcache_client
-
 $:.unshift plugin_lib, plugin_test
 
 RAILS_ENV = "test"
@@ -16,6 +14,9 @@ RAILS_ENV = "test"
 require "mocks/spawn"
 require "mocks/logger"
 require "workling"
+
+Workling.try_load_a_memcache_client
+
 require "workling/discovery"
 require "workling/starling/routing/class_and_method_routing"
 require "workling/starling/poller"